+++ lib/teamspeak3.rb
@@ -9,6 +9,7 @@ require_relative 'exceptions.rb'
 
 # library classes ordered alphabetically
 require_relative 'channel.rb'
+require_relative 'channel_collection.rb'
 require_relative 'client.rb'
 require_relative 'command_parameter.rb'
 require_relative 'virtual_server.rb'